refactor: combine mods together and update README
