+++ tests/fixtures/html-template-cleanup/closing-of-element-with-open-child-elements/expected/x/component/component.js
