+++ tests/fixtures/html-template-cleanup/closing-of-element-with-open-child-elements/input/x/component/component.html
