+++ tests/fixtures/html-template-cleanup/end-tag-without-matching-open-element/basic/expected/x/component/component.js
