+++ tests/fixtures/html-template-cleanup/end-tag-without-matching-open-element/basic/input/x/component/component.js
