+++ tests/fixtures/html-template-cleanup/end-tag-without-matching-open-element/multiple/expected/x/component/component.js
