+++ tests/fixtures/html-template-cleanup/end-tag-without-matching-open-element/multiple/input/x/component/component.js
