+++ tests/fixtures/html-template-cleanup/eof-in-element-that-can-contain-only-text/basic/expected/x/component/component.js
