+++ tests/fixtures/html-template-cleanup/eof-in-element-that-can-contain-only-text/basic/input/x/component/component.html
