--- transforms/htmlParsing/index.js
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) 2021, salesforce.com, inc.
- * All rights reserved.
- * SPDX-License-Identifier: BSD-3-Clause
- * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
- */
-export * from './closingOfElementWithOpenChildElements'
-export * from './endTagWithoutMatchingOpenElement'
-export * from './eofInElementThatCanContainOnlyText'
