+++ transforms/htmlTemplateCleanup/closingOfElementWithOpenChildElements.js
@@ -14,23 +14,16 @@ const closeChildElements = (node, errorsMap) => {
   }
 }
 
-export const closingOfElementWithOpenChildElements = async ({ templates }) => {
-  const result = {
-    overwrite: {},
-    delete: []
+export const fixClosingOfElementWithOpenChildElements = ({ ast, file, parse5Errors }, result) => {
+  const filteredErrors = parse5Errors
+    .filter(({ code }) => code === 'closing-of-element-with-open-child-elements')
+    .map(({ code, ...location }) => [location.startOffset, location])
+
+  const errorsMap = new Map(filteredErrors)
+  if (errorsMap.size) {
+    walkParse5Ast(ast, closeChildElements, errorsMap)
+    result.overwrite[file] = serializeParse5Ast(ast)
   }
-
-  templates.forEach(({ ast, file, parse5Errors }) => {
-    const filteredErrors = parse5Errors
-      .filter(({ code }) => code === 'closing-of-element-with-open-child-elements')
-      .map(({ code, ...location }) => [location.startOffset, location])
-
-    const errorsMap = new Map(filteredErrors)
-    if (errorsMap.size) {
-      walkParse5Ast(ast, closeChildElements, errorsMap)
-      result.overwrite[file] = serializeParse5Ast(ast)
-    }
-  })
 
   return result
 }
