+++ transforms/runTransform.js
@@ -10,11 +10,7 @@ import fs from 'fs/promises'
 import { walkComponents } from './walkComponents.js'
 import { shadowToLight } from './shadowToLight/index.js'
 import { syntheticToNative } from './syntheticToNative/index.js'
-import {
-  endTagWithoutMatchingOpenElement,
-  closingOfElementWithOpenChildElements,
-  eofInElementThatCanContainOnlyText
-} from './htmlParsing/index.js'
+import { htmlTemplateCleanup } from './htmlTemplateCleanup/index.js'
 
 const include = ['.js', '.ts']
 const exclude = ['__tests__', '.stories.ts']
@@ -22,9 +18,7 @@ const exclude = ['__tests__', '.stories.ts']
 const transforms = {
   'shadow-to-light': shadowToLight,
   'synthetic-to-native': syntheticToNative,
-  'end-tag-without-matching-open-element': endTagWithoutMatchingOpenElement,
-  'closing-of-element-with-open-child-elements': closingOfElementWithOpenChildElements,
-  'eof-in-element-that-can-contain-only-text': eofInElementThatCanContainOnlyText
+  'html-template-cleanup': htmlTemplateCleanup
 }
 
 export async function runTransform (dir, transformPath) {
